backend/consumer/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0
litellm==1.59.8
prometheus-client==0.21.1